@@ -1,35 +1,36 @@
 {
-  "name": "is-supported-regexp-flag",
-  "version": "1.0.1",
-  "description": "Check whether a RegExp flag is supported. Mostly useful for `y` and `u`.",
-  "license": "MIT",
-  "repository": "sindresorhus/is-supported-regexp-flag",
-  "author": {
-    "name": "Sindre Sorhus",
-    "email": "sindresorhus@gmail.com",
-    "url": "http://sindresorhus.com"
-  },
-  "engines": {
-    "node": ">=0.10.0"
-  },
-  "scripts": {
-    "test": "mocha"
-  },
-  "files": [
-    "index.js"
-  ],
-  "keywords": [
-    "regexp",
-    "regex",
-    "re",
-    "check",
-    "is",
-    "support",
-    "supported",
-    "valid",
-    "flag"
-  ],
-  "devDependencies": {
-    "mocha": "*"
-  }
+	"name": "is-supported-regexp-flag",
+	"version": "1.0.1",
+	"description": "Check whether a RegExp flag is supported. Mostly useful for `y` and `u`.",
+	"license": "MIT",
+	"repository": "sindresorhus/is-supported-regexp-flag",
+	"author": {
+		"name": "Sindre Sorhus",
+		"email": "sindresorhus@gmail.com",
+		"url": "sindresorhus.com"
+	},
+	"engines": {
+		"node": ">=6"
+	},
+	"scripts": {
+		"test": "xo && ava"
+	},
+	"files": [
+		"index.js"
+	],
+	"keywords": [
+		"regexp",
+		"regex",
+		"re",
+		"check",
+		"is",
+		"support",
+		"supported",
+		"valid",
+		"flag"
+	],
+	"devDependencies": {
+		"ava": "*",
+		"xo": "*"
+	}
 }
